+++ protocol/carbon/metricdeconstructor/identitymetricdeconstructor_test.go
@@ -3,13 +3,15 @@ package metricdeconstructor
 import (
 	"testing"
 
+	"github.com/signalfx/golib/datapoint"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestIdentityMetricDeconstructor(t *testing.T) {
 	i := &identityMetricDeconstructor{}
-	m, d, e := i.Parse("originalmetric")
+	m, mt, d, e := i.Parse("originalmetric")
 	assert.Equal(t, "originalmetric", m, "Should get metric back")
 	assert.Equal(t, 0, len(d), "Should get no dimensions")
+	assert.Equal(t, datapoint.Gauge, mt)
 	assert.Equal(t, nil, e, "Should get no errors")
 }